remove unused arg in gv_fetchfile_flags
Since gv_fetchfile_flags was created in commit d9095cec1b in 5.9.5 in 2007
the flags argument has been unused and undefined as to its meaning.
Remove the cpu instruction overhead of pushing the flags arg in the
callers of gv_fetchfile_flags by changing it to a macro that drops out
the flags argument. Only known callers are NYTProf and core's gv_fetchfile
according to cpan grep. This commits sets up future refactoring of
gv_fetchfile*.